Finishing Elementary Gates + 16-bit wrappers
I have finished all the Elementary Gates as well
as their 16-bit wrappers. I spent an embarassing
amount of time trying to figure out a way to do the
Not16.hdl with a loop or the clock. I didn't think that
16 Not calls was the best way to do it, but after talking
to some friends and looking online, it seems like that
is the intended way.