Finished CPU.hdl
Finally got the CPU.hdl working against the
two test files. I had a few misunderstandings
when it came to the ARegister, A bit (inst[12]), and
write A bit (inst[5]), so that slowed me down a lot.
But after debugging and re-reading the instructions a
few dozen times, it started to clear up.