Initial work on the Assembler
I've started working on the assembler. I've divided it
up into a few classes, but I'm not sure if that is the
best way to handle it. I'm going to keep trying it with
this approach and reevaluate down the line. If this
turns out to be more hassle than it's worth, then I will
convert it to another structure.