Implemented the Symbol Table for Assembler
I have implemented the symbol table for the assembler.
I used a separate class for it even though it is just a dictionary
with a few helper methods. If I was to do it by my own design,
I would have made it a part of the parser class.

There is nothing too special about the symbol table compared
to the rest of the program. I had misunderstood how to calculate
the addresses since I initially thought that values could not
overlap.